feat: メソッドの実装内容をインターフェースとその実装であるクラスに切り出すようにしました。 またそれをModuleに登録しました
