+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/user/renote/mute/RenoteMuteRepositoryImpl.kt
@@ -1,84 +1,36 @@
 package net.pantasystem.milktea.data.infrastructure.user.renote.mute
 
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.withContext
 import net.pantasystem.milktea.api.misskey.users.renote.mute.RenoteMuteDTO
 import net.pantasystem.milktea.common.APIError
 import net.pantasystem.milktea.common.runCancellableCatching
 import net.pantasystem.milktea.common_android.hilt.IODispatcher
 import net.pantasystem.milktea.data.infrastructure.user.renote.mute.db.RenoteMuteDao
-import net.pantasystem.milktea.data.infrastructure.user.renote.mute.db.RenoteMuteRecord
 import net.pantasystem.milktea.data.infrastructure.user.renote.mute.delegate.CreateRenoteMuteAndPushToRemoteDelegate
 import net.pantasystem.milktea.data.infrastructure.user.renote.mute.delegate.FindRenoteMuteAndUpdateMemCacheDelegate
-import net.pantasystem.milktea.model.account.GetAccount
+import net.pantasystem.milktea.data.infrastructure.user.renote.mute.delegate.SyncRenoteMuteDelegate
 import net.pantasystem.milktea.model.user.User
 import net.pantasystem.milktea.model.user.renote.mute.RenoteMute
 import net.pantasystem.milktea.model.user.renote.mute.RenoteMuteRepository
 import javax.inject.Inject
 
 internal class RenoteMuteRepositoryImpl @Inject constructor(
-    private val getAccount: GetAccount,
     private val renoteMuteDao: RenoteMuteDao,
     private val renoteMuteApiAdapter: RenoteMuteApiAdapter,
     private val isSupportRenoteMuteInstance: IsSupportRenoteMuteInstance,
-    private val unPushedRenoteMutesDiffFilter: UnPushedRenoteMutesDiffFilter,
     private val cache: RenoteMuteCache,
     private val findRenoteMuteAndUpdateMemCache: FindRenoteMuteAndUpdateMemCacheDelegate,
     private val createAndPushToRemote: CreateRenoteMuteAndPushToRemoteDelegate,
+    private val syncRenoteMuteDelegate: SyncRenoteMuteDelegate,
     @IODispatcher private val coroutineDispatcher: CoroutineDispatcher,
 ) : RenoteMuteRepository {
 
-    override suspend fun syncBy(accountId: Long): Result<Unit> = runCancellableCatching {
-        withContext(coroutineDispatcher) {
-            // NOTE: キャッシュを更新しておく
-            findBy(accountId).getOrThrow()
-
-            if (!isSupportRenoteMuteInstance(accountId)) {
-                return@withContext
-            }
-
-            val account = getAccount.get(accountId)
-
-            val mutes: List<RenoteMuteDTO> = FindAllRemoteRenoteMutes(
-                account,
-                renoteMuteApiAdapter
-            ).invoke()
-
-            val unPushedRenoteMutes = unPushedRenoteMutesDiffFilter(
-                mutes,
-                renoteMuteDao.findByUnPushed(account.accountId)
-                    .map {
-                        it.toModel()
-                    }
-            )
-
-
-            coroutineScope {
-                unPushedRenoteMutes.map {
-                    async {
-                        createAndPushToRemote(it.userId)
-                    }
-                }
-            }.awaitAll()
-
-            renoteMuteDao.deleteBy(account.accountId)
-            cache.clearBy(account.accountId)
-
-            val newModels = mutes.map {
-                it.toModel(account.accountId)
-            }
-            renoteMuteDao.insertAll(
-                newModels.map {
-                    RenoteMuteRecord.from(it)
-                }
-            )
-            cache.addAll(newModels, true)
-
-        }
-    }
+    override suspend fun syncBy(accountId: Long): Result<Unit> = syncRenoteMuteDelegate(accountId)
 
     override suspend fun syncBy(userId: User.Id): Result<Unit> = runCancellableCatching {
         withContext(coroutineDispatcher) {
